test(dropdown): temp fix for breaking test
